Always submit initial results
When a result is detected for the first time it is always submitted.
This can be blocked by the FuzzManager frequent flag. The submission
is forced if the result has only been found once (including parallel
instances). For subsequent submissions we check if the result is
frequent and report accordingly. This takes into account results found
by parallel instances.